refactor: move ConvDims/PoolDims to internal/tensor for shared access
Eliminates autodiff → cpu cross-dependency. The autodiff maxpool2d op
now imports PoolDims from internal/tensor instead of internal/backend/cpu.
CPU package keeps type aliases for backward compatibility.

Fixes #48
